Using the center of gravity version with integers now instead

+++ ino_tree2/decision_forest.h
@@ -1,1 +1,2 @@
-unsigned char evaluate_forest(float* args);
+// unsigned char evaluate_forest(float* args);
+unsigned char evaluate_forest(long* args);
